Add user id to Meal, and implement methods: save, delete, get, getAll of InMemoryMealRepositoryImpl

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,17 +1,23 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.AuthorizedUser;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
@@ -23,24 +29,41 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     public Meal save(Meal meal) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(AuthorizedUser.id());
         }
         repository.put(meal.getId(), meal);
         return meal;
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id) {
+        log.info("delete meal {}", id);
+        if (repository.containsKey(id) && repository.get(id).getUserId().equals(AuthorizedUser.id())) {
+            repository.remove(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        log.info("get meal {}", id);
+        if (repository.containsKey(id) && repository.get(id).getUserId().equals(AuthorizedUser.id())) {
+            return repository.get(id);
+        } else {
+            return null;
+        }
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        return repository
+                .values()
+                .stream()
+                .filter(v -> v.getUserId().equals(AuthorizedUser.id()))
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(Collectors.toList());
     }
 }
